// src/app/components/explorer/explorer.types.ts
export type ColumnDataType = "string" | "number" | "boolean" | "date" | "reference" | "unknown";

export type RendererId = "string-section-renderer" | "boolean-section-renderer" | "date-section-renderer" |
  "media-section-renderer" | "reference-section-renderer" | "string-object-renderer" | "boolean-object-renderer" |
  "date-object-renderer" | "reference-object-renderer" | "media-object-renderer" | "localized-string-renderer" |
  "localized-media-renderer" | string;

export interface ExplorerTarget {
  alias: string;
  target: string;
  tableName: string;
  name: LocalizedString[];
  description: LocalizedString[];
  icon: Media;
  columns: ExplorerColumn[];
  actions: ExplorerAction[];
  defaultActionCreate: boolean;
  defaultActionSave: boolean;
  defaultActionDelete: boolean;
  defaultActionDuplicate: boolean;
  canRead: UserRole[];
  canWrite: UserRole[];
  size?: number;
}

export interface ExplorerColumn {
  id: string;
  property: string;
  name: LocalizedString[];
  description: LocalizedString[];
  type: ColumnDataType | string;
  primary: boolean;
  unique: boolean;
  multiple: boolean;
  referencedTableName: string;
  referencedEntityName: string;
  sectionPriority: number;
  objectPriority: number;
  sectionEnabled: boolean;
  objectEnabled: boolean;
  sectionVisibility: boolean;
  objectVisibility: boolean;
  virtual: boolean;
  sectionRenderer: ExplorerColumnRenderer;
  objectRenderer: ExplorerColumnRenderer;
  sectionRendererParams: object;
  objectRendererParams: object;
  tab: ExplorerTab;
}

export class ExplorerAction {
  code: string;
  name: LocalizedString[];
  description: LocalizedString[];
  type: ExplorerVariation;
  params: object;
}

export interface ExplorerTabSize {
  tablet: number;
  desktop: number;
}

export interface ExplorerTab {
  id: string;
  name: LocalizedString[];
  priority: number;
  size: ExplorerTabSize;
  target: ExplorerTarget;
}

export type ExplorerVariation = "section" | "object";

export class ExplorerColumnRenderer {
  code: string;
  name: LocalizedString[];
  description: LocalizedString[];
  type: ExplorerVariation;
  params: object;
}

export interface TargetData {
  primaryColumn: ExplorerColumn;
  namedColumn: ExplorerColumn;
  entity: ExplorerTarget;
}

export interface SectionFilterDialogConfig {
  column: ExplorerColumn;
  paramsSnapshot: () => Params;
  navigate: (queryParams: Params, queryParamsHandling?: QueryParamsHandling) => void;
}

export interface SectionDialogConfig {
  target: TargetData;
  multi?: boolean;
  initialPageableParams?: PageableParams;
}

export interface ObjectDialogConfig {
  id: string;
  target: string;
}

export interface ExplorerObjectDto<T = unknown> {
  target: string;
  entity?: T;
  id?: number;
}